+++ meta/range.rb
@@ -9,9 +9,7 @@ Mutant::Meta::Example.add :irange do
   mutation '0..100'
   mutation '2..100'
   mutation 'nil..100'
-  mutation 'self..100'
   mutation '1..nil'
-  mutation '1..self'
   mutation '1..0'
   mutation '1..1'
   mutation '1..99'
@@ -28,9 +26,7 @@ Mutant::Meta::Example.add :erange do
   mutation '0...100'
   mutation '2...100'
   mutation 'nil...100'
-  mutation 'self...100'
   mutation '1...nil'
-  mutation '1...self'
   mutation '1...0'
   mutation '1...1'
   mutation '1...99'
@@ -47,7 +43,6 @@ unless RUBY_VERSION.start_with?('2.5')
     mutation '0...'
     mutation '2...'
     mutation 'nil...'
-    mutation 'self...'
   end
 
   Mutant::Meta::Example.add :irange do
@@ -58,6 +53,5 @@ unless RUBY_VERSION.start_with?('2.5')
     mutation '0..'
     mutation '2..'
     mutation 'nil..'
-    mutation 'self..'
   end
 end