+++ meta/while.rb
@@ -4,8 +4,6 @@ Mutant::Meta::Example.add :while do
   source 'while true; foo; bar; end'
 
   singleton_mutations
-  mutation 'while true; self; bar; end'
-  mutation 'while true; foo; self; end'
   mutation 'while true; bar; end'
   mutation 'while true; foo; end'
   mutation 'while true; end'